Add link to built-in emojis
